Pick up mods-architecture review follow-ups
Extract useShopEncounter (activeShop/shopStock + open/buy/sell
handlers) out of SiteMapScreen's click-handling switch — shop is one
hook call instead of five entangled pieces of state, giving future-you
a single seam to move into a mods/shop/ boundary instead of untangling
it from the screen later.

Consolidate the three separate reward-type dispatchers (SiteMapScreen's
applyReward + rareItemDisplay, ChestRewardFlow's inline switch) into
one apply-dispatcher (useApplyReward) and one display-dispatcher
(rewardText/rewardEmoji in rewardDisplay.ts). ChestRewardFlow keeps its
own richer per-type visuals (HieroglyphTile with fragment progress,
rarity) layered on top of the shared text/icon; that's UI-specific
polish, not reward-type logic, so it stays local.

Pointer comments on TreasureReward's closed union, validateEconomyGuard,
and assignPuzzleRewards — each is the ad hoc precursor to something
docs/mods-architecture.md replaces them with (currency-id registry,
dependency-graph guard, weighted-fill allocator respectively). Not
generalizing now; just breadcrumbs so the reasoning isn't rediscovered
from scratch.

No behavior change — 740 tests still pass.
